requirements.txt
Pillow = 6.2.1
matplotlib = 3.1.1
numpy = 1.17.4
opencv-python = 4.0.1
pandas = 0.25.1
argparse = 1.1
open3d = 0.8.0.0
pytesseract = 0.3.0
imutils = 0.5.3